Add script variables and source for clarity
To improve ease of maintenance, this update includes a new file .vars that
contains directory variables. These variables will make it easier to edit the
structure of the directory if changes are necessary in the future.

The .config file is now sourced in scripts 1 and 3 to directly provide the
download URLs. It is done purely in sh instead of using grep and cut.