Fix vim dependencies
vim is currently builded with the dependencies for libcanberra and libacl. But these dependencies doesn't show up in the bundles. So in order to fix it, I just changed the build to not include these dependencies. Furthermore vim needs libncursesw.so.6 in order to start. So I just changed the version of the ncurses build from 5 to 6.